updates to inventory management
trying to make it so that items can be dropped and picked back up, had to comment out some of the code for now, but I'll figure it out eventually.  going to want to be able to stack similar items at some point too, but that's down the road a ways.  player inventory is now emptied upon player death.