hotbar upgrade
hotbar is more dynamic, when player uses an item, it uses the item, no longer pops from back of vector.  when an item is added to the player's hotbar a check to make sure the bar isn't full is done, and a check for the first emptySlot is done to determine which index to insert the item.